--- src/main/java/parser/ProjectParserHelper.java
@@ -1,18 +1,0 @@
-package parser;
-
-import model.tree.edge.RelationshipIdentifier;
-import model.tree.node.LeafNode;
-import model.tree.node.PackageNode;
-
-import java.nio.file.Path;
-import java.util.Map;
-
-public interface ProjectParserHelper {
-
-    LeafNode createLeafNode(Path path);
-
-    RelationshipIdentifier createRelationshipIdentifier(Map<Path, PackageNode> packageNodes);
-
-    FileVisitor createFileVisitor();
-
-}
