+++ src/test/java/parser/javaparser/TreeStructureArchitectureTest.java
@@ -8,7 +8,7 @@ import model.tree.node.PackageNode;
 import org.junit.jupiter.api.Test;
 import parser.Parser;
 import parser.ParserType;
-import parser.ProjectParser;
+import parser.ProjectParserFactory;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -25,7 +25,9 @@ public class TreeStructureArchitectureTest {
 
     @Test
     void getFieldAndMethodTypesTest() throws IOException {
-        Parser parser = new ProjectParser(parserType);
+        ProjectParserFactory projectParserFactory = new ProjectParserFactory(parserType);
+        Parser parser = projectParserFactory.createProjectParser();
+
         parser.parseSourcePackage(Paths.get(currentDirectory.toRealPath() + "\\src\\test\\resources\\LatexEditor\\src"));
         Map<Path, PackageNode> packages = parser.getPackageNodes();
         List<String> methodReturnTypes = new ArrayList<>(Arrays.asList("Constructor", "void"));
@@ -60,7 +62,9 @@ public class TreeStructureArchitectureTest {
 
     @Test
     void leafNodeRelationshipsTest() throws IOException {
-        Parser parser = new ProjectParser(parserType);
+        ProjectParserFactory projectParserFactory = new ProjectParserFactory(parserType);
+        Parser parser = projectParserFactory.createProjectParser();
+
         parser.parseSourcePackage(Paths.get(currentDirectory.toRealPath() + "\\src\\test\\resources\\LatexEditor\\src"));
         Map<Path, PackageNode> packages = parser.getPackageNodes();
         PackageNode commandPackage = packages.get(Paths.get(currentDirectory.toRealPath().normalize().toString(), "\\src\\test\\resources\\LatexEditor\\src\\controller\\commands"));
@@ -122,7 +126,9 @@ public class TreeStructureArchitectureTest {
 
     @Test
     void leafNodeInheritanceRelationshipTest() throws IOException {
-        Parser parser = new ProjectParser(parserType);
+        ProjectParserFactory projectParserFactory = new ProjectParserFactory(parserType);
+        Parser parser = projectParserFactory.createProjectParser();
+
         parser.parseSourcePackage(Paths.get(currentDirectory.toRealPath() + "\\src\\test\\resources\\ParserTesting"));
         Map<Path, PackageNode> packages = parser.getPackageNodes();
         PackageNode sourcePackage = packages.get(Paths.get(currentDirectory.toRealPath().normalize().toString(), "\\src\\test\\resources\\ParserTesting"));
@@ -155,7 +161,9 @@ public class TreeStructureArchitectureTest {
 
     @Test
     void packageNodeRelationshipsTest() throws IOException {
-        Parser parser = new ProjectParser(parserType);
+        ProjectParserFactory projectParserFactory = new ProjectParserFactory(parserType);
+        Parser parser = projectParserFactory.createProjectParser();
+
         parser.parseSourcePackage(Paths.get(currentDirectory.toRealPath() + "\\src\\test\\resources\\LatexEditor\\src"));
         Map<Path, PackageNode> packages = parser.getPackageNodes();
 
@@ -203,7 +211,9 @@ public class TreeStructureArchitectureTest {
 
     @Test
     void leafNodeTypesTest() throws IOException {
-        Parser parser = new ProjectParser(parserType);
+        ProjectParserFactory projectParserFactory = new ProjectParserFactory(parserType);
+        Parser parser = projectParserFactory.createProjectParser();
+
         parser.parseSourcePackage(Paths.get(currentDirectory.toRealPath() + "\\src\\test\\resources\\ParserTesting"));
         Map<Path, PackageNode> packages = parser.getPackageNodes();
         PackageNode sourcePackage = packages.get(Paths.get(currentDirectory.toRealPath().normalize().toString(), "\\src\\test\\resources\\ParserTesting"));
