+++ src/test/java/parser/jdt/NodeParsingTest.java
@@ -6,7 +6,7 @@ import model.tree.node.PackageNode;
 import org.junit.jupiter.api.Test;
 import parser.Parser;
 import parser.ParserType;
-import parser.ProjectParser;
+import parser.ProjectParserFactory;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -23,7 +23,9 @@ class NodeParsingTest {
 
 	@Test
 	void methodReturnTypesTest() throws IOException {
-		Parser parser = new ProjectParser(parserType);
+		ProjectParserFactory projectParserFactory = new ProjectParserFactory(parserType);
+		Parser parser = projectParserFactory.createProjectParser();
+
 		parser.parseSourcePackage(Paths.get(currentDirectory.toRealPath() + "\\src\\test\\resources\\LatexEditor\\src"));
 		Map<Path, PackageNode> packages = parser.getPackageNodes();
 		PackageNode commandPackage = packages.get(Paths.get(currentDirectory.toRealPath().normalize().toString(),
@@ -42,7 +44,9 @@ class NodeParsingTest {
 
 	@Test
 	void methodParameterTypesTest() throws IOException {
-		Parser parser = new ProjectParser(parserType);
+		ProjectParserFactory projectParserFactory = new ProjectParserFactory(parserType);
+		Parser parser = projectParserFactory.createProjectParser();
+
 		parser.parseSourcePackage(Paths.get(currentDirectory.toRealPath() + "\\src\\test\\resources\\LatexEditor\\src"));
 		Map<Path, PackageNode> packages = parser.getPackageNodes();
 		PackageNode commandPackage = packages.get(Paths.get(currentDirectory.toRealPath().normalize().toString(),
@@ -61,7 +65,9 @@ class NodeParsingTest {
 
 	@Test
 	void fieldTypesTest() throws IOException {
-		Parser parser = new ProjectParser(parserType);
+		ProjectParserFactory projectParserFactory = new ProjectParserFactory(parserType);
+		Parser parser = projectParserFactory.createProjectParser();
+
 		parser.parseSourcePackage(Paths.get(currentDirectory.toRealPath() + "\\src\\test\\resources\\LatexEditor\\src"));
 		Map<Path, PackageNode> packages = parser.getPackageNodes();
 		PackageNode commandPackage = packages.get(Paths.get(currentDirectory.toRealPath().normalize().toString(),
@@ -80,7 +86,9 @@ class NodeParsingTest {
 
 	@Test
 	void leafNodeTypesTest() throws IOException {
-		Parser parser = new ProjectParser(parserType);
+		ProjectParserFactory projectParserFactory = new ProjectParserFactory(parserType);
+		Parser parser = projectParserFactory.createProjectParser();
+
 		parser.parseSourcePackage(Paths.get(currentDirectory.toRealPath() + "\\src\\test\\resources\\ParserTesting"));
 		Map<Path, PackageNode> packages = parser.getPackageNodes();
 		PackageNode inheritancePackage = packages.get(Paths.get(currentDirectory.toRealPath().normalize().toString(), "\\src\\test\\resources\\ParserTesting"));
